moved app to webflux reactive and tested in local

+++ client-app/src/main/java/com/streaming/configuration/properties/model/MetricsConfigurationProperties.java
@@ -24,7 +24,7 @@ public class MetricsConfigurationProperties {
 
     @Data
     public static class Threshold {
-        private int highCpuPercentage;
+        private double highCpuPercentage;
     }
 
     @Data
